+++ LibraWallet/Sections/HomeSection/ViolasTransfer/Controller/ViolasTransferViewController.swift
@@ -104,7 +104,7 @@ extension ViolasTransferViewController: ViolasTransferViewDelegate {
         vc.hidesBottomBarWhenPushed = true
         self.navigationController?.pushViewController(vc, animated: true)
     }
-    func confirmTransfer(amount: Double, address: String, fee: Double) {
+    func confirmTransfer(amount: UInt64, address: String, fee: UInt64) {
         WalletManager.unlockWallet(controller: self, successful: { [weak self] (mnemonic) in
             self?.detailView.toastView?.show(tag: 99)
             self?.dataModel.sendViolasTransaction(sendAddress: self?.wallet?.tokenAddress ?? "",
